moved authContext to context folder
