+++ client/app/screens/TrailScreen.tsx
@@ -2,7 +2,7 @@ import { useState, useEffect, useContext } from "react";
 import { StackNativeScreenProps } from "../interfaces/StackParamList";
 import * as TaskManager from "expo-task-manager";
 import * as Location from "expo-location";
-import { AuthContext } from "../utils/authContext";
+import { AuthContext } from "../context/authContext";
 import { checkStatus } from "../utils/permissionHelpers";
 
 // Components
